ndctl/list: Add --debug option
Add an flag to make 'ndctl list` print the libndctl debug output to stderr.
Ideally this would be consistent with ndctl-create-namespace which uses -v
to enable debug output, but -v is already used by ndctl-list to set the
level of detail in the JSON output.

We could add another -v level, but making that enable debug output seemed
like it would just be annoying for script writers.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>