ndctl/namespace: Check for seed namespaces earlier
When creating an fsdax or devdax namespace we need to verify that the
seed namespaces exist. This patch reworks the validation so that it's
done earlier to simplify the subsequent patches in the series.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
No functional changes, I hope.

v3: Don't fail early if the user specifies fsdax mode without pfn
    support. Instead, if we try to validate a feature that requires
    pfn support, fail the validation at that point.