ndctl/namespace: Choose a default alignment from the seed NS
When creating a mem or dax namespace ndctl uses a default alignment of 2MB
when the user does not explicitly supply one. This works on most systems (x86,
ARM, PPC64 with radix MMU), but it fails when the kernel does not support a
2MB page size (PPC64 with hash MMU).

This patch makes ndctl use the alignement of the relevant seed namespace
as the default instead. The kernel ensures the seeds always have a sensible
alignment for the system they are running on so this technique should be
more portable.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>